requirements.txt
asgiref==3.4.1
boto3==1.21.37
botocore==1.24.37

cryptography==36.0.1
Django==3.2.11
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==4.4.0

jwt==1.3.1
mysqlclient==2.1.0
Pillow==8.4.0

PyJWT==1.7.1

pytz==2021.3
s3transfer==0.5.2
simplejwt==2.0.1
six==1.16.0
sqlparse==0.4.2
typing==3.7.4.3

urllib3==1.26.9
whitenoise==5.3.0